# internal/services/notification_service/domain/props/send_email.go
package props

type SendEmailRequest struct {
	Subject string   `json:"subject"`
	To      []string `json:"to"`
	Data    string   `json:"data"`
}

type SendEmailResp struct {
	Status string `json:"status"`
}