# internal/services/user_service/docs/swagger.yaml
basePath: /
definitions:
  k071123_internal_services_notification_service_domain_props.SendEmailRequest:
    properties:
      data:
        type: string
      subject:
        type: string
      to:
        items:
          type: string
        type: array
    type: object
  k071123_internal_services_notification_service_domain_props.SendEmailResp:
    properties:
      status:
        type: string
    type: object
  k071123_internal_services_user_service_domain_models.User:
    properties:
      birth_date:
        type: string
      email:
        type: string
      first_name:
        type: string
      phone_number:
        type: string
      role:
        $ref: '#/definitions/k071123_internal_shared_permissions.Role'
      second_name:
        type: string
      status:
        $ref: '#/definitions/k071123_internal_services_user_service_domain_models_user_status.Status'
      uuid:
        type: string
    type: object
  k071123_internal_services_user_service_domain_models_user_status.Status:
    enum:
    - active
    - inactive
    type: string
    x-enum-varnames:
    - Active
    - Inactive
  k071123_internal_services_user_service_domain_props.ConfirmCodeReq:
    properties:
      code:
        type: string
      email:
        type: string
    type: object
  k071123_internal_services_user_service_domain_props.ConfirmCodeResp:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  k071123_internal_services_user_service_domain_props.CreateUserReq:
    properties:
      birth_date:
        type: string
      first_name:
        type: string
      phone_number:
        type: string
      second_name:
        type: string
      status:
        $ref: '#/definitions/k071123_internal_services_user_service_domain_models_user_status.Status'
    type: object
  k071123_internal_services_user_service_domain_props.CreateUserResp:
    properties:
      user:
        $ref: '#/definitions/k071123_internal_services_user_service_domain_models.User'
    type: object
  k071123_internal_services_user_service_domain_props.SendCodeReq:
    properties:
      email:
        type: string
    type: object
  k071123_internal_services_user_service_domain_props.SendCodeResponse:
    properties:
      status:
        type: string
    type: object
  k071123_internal_shared_permissions.Role:
    enum:
    - superadmin
    - parkingadmin
    - user
    type: string
    x-enum-varnames:
    - SuperAdmin
    - ParkingAdmin
    - Default
  k071123_internal_utils_errs.Code:
    enum:
    - 500
    - 400
    - 401
    - 403
    - 404
    - 405
    - 409
    - 422
    - 429
    - 503
    - 504
    type: integer
    x-enum-comments:
      BadRequest: Некорректный запрос
      Conflict: Конфликт (например, дублирующая запись)
      Forbidden: Запрещено
      InternalServerError: Внутренняя ошибка сервера
      MethodNotAllowed: Метод не разрешён
      NotFound: Не найдено
      ServiceUnavailable: Сервис временно недоступен
      Timeout: Таймаут запроса
      TooManyRequests: Слишком много запросов
      Unauthorized: Неавторизованный доступ
      UnprocessableEntity: Невозможность обработать запрос (например, валидация)
    x-enum-varnames:
    - InternalServerError
    - BadRequest
    - Unauthorized
    - Forbidden
    - NotFound
    - MethodNotAllowed
    - Conflict
    - UnprocessableEntity
    - TooManyRequests
    - ServiceUnavailable
    - Timeout
  k071123_internal_utils_errs.Error:
    properties:
      code:
        $ref: '#/definitions/k071123_internal_utils_errs.Code'
      details:
        type: string
      message:
        type: string
    type: object
host: localhost:7800
info:
  contact: {}
  description: API для работы с пользователем
  title: User Service API
  version: "2.0"
paths:
  /api/v1/auth/confirm_code:
    post:
      consumes:
      - application/json
      description: Подтвердить отправленный код
      parameters:
      - description: Код потверждения и почта.
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/k071123_internal_services_user_service_domain_props.ConfirmCodeReq'
      produces:
      - application/json
      responses:
        "200":
          description: Успешная авторизация.
          schema:
            $ref: '#/definitions/k071123_internal_services_user_service_domain_props.ConfirmCodeResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/k071123_internal_utils_errs.Error'
        "404":
          description: Profile not found
          schema:
            $ref: '#/definitions/k071123_internal_utils_errs.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/k071123_internal_utils_errs.Error'
      security:
      - BearerAuth: []
      summary: Подтвердить отправленный код
      tags:
      - Auth
  /api/v1/auth/send_code:
    post:
      consumes:
      - application/json
      description: Отправить код подтверждения
      parameters:
      - description: Почта пользователя.
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/k071123_internal_services_user_service_domain_props.SendCodeReq'
      produces:
      - application/json
      responses:
        "200":
          description: Статус отправки Email.
          schema:
            $ref: '#/definitions/k071123_internal_services_user_service_domain_props.SendCodeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/k071123_internal_utils_errs.Error'
        "404":
          description: Profile not found
          schema:
            $ref: '#/definitions/k071123_internal_utils_errs.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/k071123_internal_utils_errs.Error'
      security:
      - BearerAuth: []
      summary: Отправить код подтверждения
      tags:
      - Auth
  /api/v1/sender/email:
    post:
      consumes:
      - application/json
      description: Отправить Email
      parameters:
      - description: Данные для отправки email.
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/k071123_internal_services_notification_service_domain_props.SendEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Статус отправленного email.
          schema:
            $ref: '#/definitions/k071123_internal_services_notification_service_domain_props.SendEmailResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/k071123_internal_utils_errs.Error'
        "404":
          description: Profile not found
          schema:
            $ref: '#/definitions/k071123_internal_utils_errs.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/k071123_internal_utils_errs.Error'
      summary: Отправить Email
      tags:
      - EmailSender
  /api/v1/users:
    post:
      consumes:
      - application/json
      description: Создать пользователя
      parameters:
      - description: Данные для создания пользователя.
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/k071123_internal_services_user_service_domain_props.CreateUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ. Модель пользователя.
          schema:
            $ref: '#/definitions/k071123_internal_services_user_service_domain_props.CreateUserResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/k071123_internal_utils_errs.Error'
        "404":
          description: Profile not found
          schema:
            $ref: '#/definitions/k071123_internal_utils_errs.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/k071123_internal_utils_errs.Error'
      security:
      - BearerAuth: []
      summary: Создать пользователя
      tags:
      - User
securityDefinitions:
  BearerAuth:
    description: '"Enter your Bearer token in the format: `Bearer {token}`"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"